+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - Put your changes here...
 
+## 1.0.2
+
+- Fixed bug causing custom `res` variables to not be set properly during certain default render method calls.
+- Fixed broken Express 5 support.
+- Updated dependencies.
+
 ## 1.0.1
 
 - Fixed broken postinstall script.